Files added and moved after rebase of this branch.
