+++ src/main/java/use_cases/reload_game_use_case/ReloadGameDsResponseModel.java
@@ -1,6 +1,6 @@
 package use_cases.reload_game_use_case;
 
-import CoreEntities.Player.Player;
+import core_entities.player.Player;
 import core_entities.game_parts.Board;
 
 public class ReloadGameDsResponseModel {
