+++ src/test/java/io/display_logic/ViewTest.java
@@ -1,8 +1,9 @@
-package CoreEntities.IO;
+package io.display_logic;
+
+import io.windows.HomeWindow;
+import io.display_logic.View;
 
 import javax.swing.*;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class ViewTest {
     public static void main(String[] args) {
