+++ src/test/java/place_word_refill_user_story/PlaceWordRefillInteractorTest.java
@@ -1,13 +1,6 @@
 package place_word_refill_user_story;
 
-import CoreEntities.Player.ComputerPlayer;
-import CoreEntities.Player.HumanPlayer;
-import core_entities.game_parts.Bag;
-import core_entities.game_parts.GameState;
-import core_entities.game_parts.LetterRack;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class PlaceWordRefillInteractorTest {
